ScanRipsManager and ScanMoviesManager now inherit from GeneralScanManager
